chore(registration-api): Move Enhet and FramedCatalog to model package
